feat: pure js crypto
