+++ package.json
@@ -40,16 +40,21 @@
 		"lib/*"
 	],
 	"dependencies": {
+		"@noble/ciphers": "^1.3.0",
+		"@noble/curves": "^1.9.6",
+		"@noble/hashes": "^1.8.0",
 		"@peculiar/asn1-ecc": "^2.3.14",
 		"@peculiar/asn1-schema": "^2.3.13",
 		"@peculiar/x509": "^1.12.3",
-		"@stablelib/chacha20poly1305": "^1.0.1"
+		"@stablelib/chacha20poly1305": "^1.0.1",
+		"micro-rsa-dsa-dh": "^0.1.0"
 	},
 	"devDependencies": {
 		"@adiwajshing/eslint-config": "github:adiwajshing/eslint-config",
 		"@commitlint/cli": "^17.8.1",
 		"@commitlint/config-conventional": "^17.8.1",
 		"@swc-node/register": "^1.10.9",
+		"@swc/jest": "^0.2.39",
 		"@types/chance": "^1.1.6",
 		"@types/jest": "^29.5.14",
 		"@types/node": "^20.17.9",
@@ -59,7 +64,6 @@
 		"eslint": "^8.57.1",
 		"jest": "^29.7.0",
 		"pino": "^9.5.0",
-		"ts-jest": "^29.2.5",
 		"ts-node": "^10.9.2",
 		"typescript": "^4.9.5"
 	}
