+++ src/utils/decryption-utils.ts
@@ -1,7 +1,7 @@
 import { crypto } from '../crypto'
 import { CipherSuite, HashAlgorithm } from '../types'
 import { SUPPORTED_CIPHER_SUITE_MAP } from './constants'
-import { concatenateUint8Arrays, isSymmetricCipher, strToUint8Array, uint8ArrayToDataView } from './generics'
+import { concatenateUint8Arrays, hkdfExpand, isSymmetricCipher, strToUint8Array, uint8ArrayToDataView } from './generics'
 import { packWithLength } from './packets'
 
 type DeriveTrafficKeysOptions = {
@@ -257,7 +257,7 @@ export async function hkdfExtractAndExpandLabel(algorithm: HashAlgorithm, secret
 	])
 
 	const key = await crypto.importKey(algorithm, secret)
-	return crypto.expand(algorithm, length, key, length, hkdfLabel)
+	return hkdfExpand(algorithm, length, key, length, hkdfLabel, crypto)
 }
 
 export async function getHash(msgs: Uint8Array[] | Uint8Array, cipherSuite: CipherSuite) {
